introduce report.* to manage reports.
