+++ internal/smash/structs.go
@@ -18,21 +18,3 @@ type App struct {
 	Args      []string
 	Locations []string
 }
-type RunSummary struct {
-	DuplicateFileSizeF string
-	DuplicateFileSize  uint64
-	TotalFiles         int64
-	TotalFileErrors    int64
-	ElapsedTime        int64
-	UniqueFiles        int64
-	EmptyFiles         int64
-	DuplicateFiles     int64
-}
-type SmashFile struct {
-	Filename    string
-	Hash        string
-	FileSizeF   string
-	FileSize    uint64
-	ElapsedTime int64
-	FullHash    bool
-}
